+++ src/py42/clients/__init__.py
@@ -12,5 +12,7 @@ Clients = namedtuple(
         "cases",
         "loginconfig",
         "trustedactivities",
+        "userriskprofile",
+        "watchlists",
     ],
 )